moved methods to models/utils.py
